[TESTS][JL] Restructure folder based on storybook
